fix type of m_size & m_addr to match prototypes 9from Ralph)
SCCS-vsn: 8.3

+++ sys/sys/map.h
@@ -35,7 +35,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)map.h	8.2 (Berkeley) 01/21/94
+ *	@(#)map.h	8.3 (Berkeley) 01/26/94
  */
 
 /*
@@ -67,8 +67,8 @@ struct map {
 };
 
 struct mapent {
-	int	m_size;			/* size of this segment of the map */
-	int	m_addr;			/* start of segment */
+	long	m_size;			/* size of this segment of the map */
+	long	m_addr;			/* start of segment */
 };
 
 #ifdef KERNEL
